requirements.txt
langchain
langchain-qdrant
langchain_community
langchain_ollama
langchain_huggingface
langchain_text_splitters
langchain_groq
pypdf
python-dotenv
streamlit
ipykernel
mlflow
mlflow[langchain]
huggingface
sentence_transformers
fastapi[standard]
pydantic
ragas
psycopg2-binary
